+++ sdk/agentserver/azure-ai-agentserver-activity/CHANGELOG.md
@@ -20,5 +20,4 @@
 
 ### Other Changes
 
-- Requires `azure-ai-agentserver-core>=2.0.0b7`.
-
+- Requires `azure-ai-agentserver-core>=2.0.0b8`.